Add node and express to project. Modify structure to place project and index in proper folders.
