+++ package.json
@@ -1,0 +1,18 @@
+{
+  "name": "portfolio-mean",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "start": "node ./bin/www"
+  },
+  "dependencies": {
+    "body-parser": "~1.12.0",
+    "cookie-parser": "~1.3.4",
+    "debug": "~2.1.1",
+    "ejs": "~2.3.1",
+    "express": "~4.12.2",
+    "mongoose": "^4.0.2",
+    "morgan": "~1.5.1",
+    "serve-favicon": "~2.2.0"
+  }
+}
